Fix data updates for accessors using the DAQ trigger.
Those accesors were updated by the DAQ trigger at the same time the DAQ
was updated. Since so far we only waited for the trigger to be updated
we missed the update of those accessors. Now we wait for the DAQ trigger
and all accessors that use the DAQ trigger as external trigger.